+++ NEWS.md
@@ -3,6 +3,7 @@
 Updates:
 
 - Fixed [package-breaking user agent issue](https://github.com/andreweatherman/toRvik/issues/1) that threw HTTP 403 error on Windows machines by using `withr`
+- Fixed [function-breaking file issue](https://github.com/andreweatherman/toRvik/issues/3) that caused `bart_player_game` to fail
 - Added [`bart_team_box`](https://www.torvik.dev/reference/bart_team_box.html) to return team box statistics and per-game averages  
 - Added argument to filter `bart_transfers` by [active portal players and committed ones](https://github.com/andreweatherman/toRvik/issues/2)